grist-deployment-tests: Introduce NO_CLEANUP env var for e2e tests

+++ grist-deployment-tests/test/e2e/utils.js
@@ -1,5 +1,7 @@
 import { browser } from "@wdio/globals";
 
+const noCleanup = ["1", "y", "yes", "true"].includes(process.env.NO_CLEANUP);
+
 // This id is a virtual one that is meant to target logged in user's personnal org.
 // It remains constant for whoever calls the API and on whatever Grist instance.
 const PERSONAL_ORG_ID = 0;
@@ -41,7 +43,9 @@ export async function withTmpWorkspace(callback) {
     await $(`=${wsName}`).click();
     await callback(wsId, wsName);
   } finally {
-    await deleteWorkspace(wsId);
+    if (!noCleanup) {
+      await deleteWorkspace(wsId);
+    }
   }
 }
